Change everything to double

+++ lib/model/song.dart
@@ -7,7 +7,7 @@ class Song {
   final String name;
   final String location;
   final double bpm;
-  final int offset;
+  final double offset;
 
   Song({
     required this.id,
@@ -25,7 +25,7 @@ class Song {
       name: json['name'] as String,
       location: json['location'] as String,
       bpm: bpm.toDouble(),
-      offset: offset.toInt(),
+      offset: offset.toDouble(),
     );
   }
 }
@@ -46,7 +46,7 @@ class SongsProvider with ChangeNotifier {
   String get name => songs[_selected].name;
   String get location => songs[_selected].location;
   double get bpm => songs[_selected].bpm;
-  int get offset => songs[_selected].offset;
+  double get offset => songs[_selected].offset;
 
   SongsProvider() {
     fetch();
@@ -93,7 +93,7 @@ class SongsProvider with ChangeNotifier {
           name: name,
           location: location,
           bpm: bpm.toDouble(),
-          offset: offset.toInt(),
+          offset: offset.toDouble(),
         );
         notifyListeners();
       } else {
@@ -121,7 +121,7 @@ class SongsProvider with ChangeNotifier {
           name: name,
           location: location,
           bpm: bpm.toDouble(),
-          offset: offset.toInt(),
+          offset: offset.toDouble(),
         );
         notifyListeners();
       } else {
@@ -151,7 +151,7 @@ class SongsProvider with ChangeNotifier {
           name: name,
           location: location,
           bpm: bpm.toDouble(),
-          offset: offset.toInt(),
+          offset: offset.toDouble(),
         );
         notifyListeners();
       } else {
